user actions moved to one auth actions file to avoid code repitition
