--- src/data/actions/UserActions/resetAuthResponse.js
@@ -1,5 +1,0 @@
-import RESET_AUTH_RESPONSE from '../../constants/UserConstants/resetAuthResponse'
-
-export const resetAuthResponse = () => ({
-    type: RESET_AUTH_RESPONSE
-})
