+++ src/data/actions/index.js
@@ -4,8 +4,5 @@ export * from './JobActions/editJobActions'
 export * from './JobActions/viewJobActions'
 export * from './JobActions/addJobActions'
 export * from './JobActions/deleteJobActions'
-// user actions
-export * from './UserActions/editUserActions'
-export * from './UserActions/loginUserActions'
-export * from './UserActions/logoutUserActions'
-export * from './UserActions/registerUserActions'
+// auth actions
+export * from './AuthActions'
